[BREAKING] Reverse return values for login()
Although this is a breaking change, it follows standard Perl convention,
and allows us to write clear code like:

    $bot->login($user, $pass) or croak "Login failed";

Add a private sub _is_loggedin() which uses the API to verify that
we're logged in, and logged in as the right user. login() now stores
the username you wanted to log in as. Update t/01_login.t accordingly.
Added an optional $paranoia to login() - callers encountering
difficulties with login can set this to have additional certainty about
whether their login was successful. This is intended for debugging use.

Added logout()