Issue 33: Method of mapping wiki names to domains (enwiki => en.wikipedia.org)
This adds db_to_domain() to get the domain name from a wiki/database name.
domain_to_db() performs the reverse operation. This will allow us to have
scripts that accept either --wiki=enwikibooks or --wiki=en.wikibooks.org
and know what both of those mean. Obviously, this is only worth using on
WMF wikis, since Extension:SiteMatrix is only worth using there.

Note that this currently saves a transformed hash in the bot object
($self->{'sitematrix'}) to avoid making repeated queries and re-transforming
the data (API data is a huge mess and needs lots of postprocessing to
have a sane data structure). In the future, this could be enhanced by
using Storable to save to disk. If mtime on the file is <14d, load from
file instead of re-querying the API over network, and re-transforming
the hash. I haven't done any tests to see whether this would be
worthwhile.

Misc:
    * Add t/27_sitematrix.t
    * Update MANIFEST