Issue 42: get_all_pages_in_category does not recurse on languages other than English
This does a query to get namespace names, and uses the localized name for
the category namespace to find subcategories. The query response is saved
in $self->{'ns_data'} to avoid repeated queries, which quickly gets *very*
expensive.

This also prepends the category name to avoid cminvalidcategory when callers
forget to include the namespace name.