Protect against infinite recursion in get_all_pages_in_category()
This uses a bare block to create a new lexical scope around the
sub get_all_pages_in_category(). This allow us to create %data as
a pseudo-state variable (ie we don't need Perl 5.10 for the state
pragma). So, we can check between calls to the sub whether we've
seen a category before and not recurse into it again if so. We
pass a parameter for recursive calls to the sub to avoid emptying
%data. External callers *must never* set this parameter, or they
may get incorrect data and a trout in the face from this developer.

Thanks to buu, anno, and LeoNerd from #perl on irc.freenode.net

Resolves Issue 84: get_all_pages_in_category() has no protection against infinite loops