Filter out private, fishbowl, and closed wikis in _get_sitematrix()
Since the obvious use-case for _get_sitematrix() is to get a list
of wikis to iterate over, we should prune out private and fishbowl
wikis, which you almost certainly can't edit. The data to do this
is present in the sitematrix API results, we just had to use it.

We add a hit to http://noc.wikimedia.org/conf/closed.dblist and
filter out the corresponding wikis.

This also waits until after filtering to add the domain->db
mapping, which probably improves performance and memory
footprint by a miniscule amount.

Resolves Issue 88: _get_sitematrix() should filter out closed/private/fishbowl wikis