Request the maximum number of results for best performance
Because we didn't set cmlimit, the default of 10 was used, resulting
in an order of magnitude more network round-trips than was required.
The limit could be 500 for normal users, or 5000 for those with
apihighlimits. We use 'max' here to ensure that max is always used.
This is inconsistent with other list methods, but it probably a
better way of doing things.