Rewrite test suite to use test.wikipedia.org
This rewrite uses test.wikipedia.org instead of en.wikipedia.org.
It also re-orders tests to be a bit more sensible - for example,
testing API error handling happens first, and we test fetching
page text prior to editing (which relies on fetching the text
correctly).

I also removed a special case in count_contributions() which
appeared to be pointless, and was causing failures in the test
suite. I believe it is a bug.