Fix login tests
The test account we were using was locked with CentralAuth so the
password could be distributed without concern. At the time, CA
locking had changed behaviour from not letting the user log in at
all to allowing them to log in, but restricting editing etc. That
behaviour switched back, so we cannot provide default account
credentials for an unlocked account.

To test logging in, provide the username and password of an account
in the PWPUsername and PWPPassword environment variables. This will
also enable editing tests to use your credentials.