Handle namespace #0 properly
In many places, the namespace number was treated as a boolean,
but namespace zero is a legitimate value. This bug would have
resulted in a desired namespace filter being omitted, and the
API results would include results from all namespaces instead
of only the requested namespace.

Thanks to Seth for reporting this.

Fixes GH #43