Return failure code properly in upload(); require GD for upload test
When the user omitted both data and filename when calling upload,
the error data was correctly set, but the method failed to return
prematurely, disguising the error. Instead of being informed that
the upload method requires either a filename or the binary data,
the error informed the caller that the file they attempted to
upload was empty. Now the correct error is given.