Refactor extraction of namespace aliases
This was always pretty hairy, because the response is split between
the "namespacealiases" and "namespaces" sections, and you have to
merge them.

This replaces the loop with a doubple-map that may be clearer,
because this shows the two steps separately. n is small enough (~4
for most wikis) that we don't care about efficiency. Comments have
also been added.

Addresses gh-64 in part.