Catch when a CAPTCHA has been encountered
This doesn't give you any way to deal with the CAPTCHA yet, but
it can let you know what happened. Since CAPTCHAs aren't a "real"
error condition, they're presented differently, and we didn't
understand that, leaving no error information in the Bot object's
error attribute. Now, we add a specific error constant and provide
the data needed to proceed in the details. The next step is to
augment the edit method with a way to accept both the ID and the
solution, and retry properly. It might also be interesting to try
an automatic interactive retry option.