Let devs solve CAPTCHAs if needed
edit will now return false but use a specific error code, ERR_CAPTCHA,
when a CAPTCHA has been encountered. The error details begin with
'captcharequired', and a special 'captcha' key is added with data
regarding the CAPTCHA which must be solved.

After interaction with a human, the edit can be retried, adding
the new captcha_id and captcha_solution parameters (in the
new hashref call style).

Fixes gh-2