Skip t/36-email.t
Need to figure out a new testing strategy here. [[User:Perlwikibot testing]]
was created with a confirmed email so you could send emails to it. The
account was then locked (in CentralAuth), but this still permitted emails
to be sent. MediaWiki no longer allows this. We need to figure out another
plan.