feat: get_id returns -1 now on non-existing pages
in the past get_id returned undef on errors or if a page did not exist.
there was no possibility to distinguish thos cases. this problem is
solved now.

fixes #93 (second part of two parts)

BREAKING CHANGE: get_id will return undef only on errors. a non-existing
page is not an error.